+++ src/connectionFromMongoCursor.js
@@ -1,25 +1,22 @@
-'use strict';
-
-var utils = require('./utils');
-var getOffsetsFromArgs = utils.getOffsetsFromArgs;
-var getConnectionFromSlice = utils.getConnectionFromSlice;
+const { getOffsetsFromArgs, getConnectionFromSlice } = require('./utils');
 
 /**
  * Accepts a mongodb cursor and connection arguments, and returns a connection
  * object for use in GraphQL. It uses array offsets as pagination, so pagiantion
  * will work only if the data set is satic.
  */
-function connectionFromMongoCursor(inMongoCursor, inArgs, mapper) {
-  var args = inArgs || {};
-  var mongodbCursor = inMongoCursor.clone();
+
+const connectionFromMongoCursor = (inMongoCursor, inArgs, mapper) => {
+  const args = inArgs || {};
+  const mongodbCursor = inMongoCursor.clone();
 
   // Because getting count from cursor in sharded DB is very slow
-  var collectionName = mongodbCursor.cmd.find.replace(/^[^.]+\./, '');
-  var query = mongodbCursor.cmd.query;
-  var countPromise = mongodbCursor.options.db.collection(collectionName).count(query);
+  const collectionName = mongodbCursor.cmd.find.replace(/^[^.]+\./, '');
+  const { query } = mongodbCursor.cmd;
+  const countPromise = mongodbCursor.options.db.collection(collectionName).countDocuments(query);
 
-  return countPromise.then(function fulfilled(count) {
-    var pagination = getOffsetsFromArgs(args, count);
+  return countPromise.then((count) => {
+    const pagination = getOffsetsFromArgs(args, count);
 
     // Short circuit if limit is 0; in that case, mongodb doesn't limit at all
     if (pagination.limit === 0) {
@@ -30,10 +27,10 @@ function connectionFromMongoCursor(inMongoCursor, inArgs, mapper) {
     mongodbCursor.skip(pagination.skip);
     mongodbCursor.limit(pagination.limit);
 
-    return mongodbCursor.toArray().then(function fromSlice(slice) {
+    return mongodbCursor.toArray().then((slice) => {
       return getConnectionFromSlice(slice, mapper, args, count);
     });
   });
-}
+};
 
 module.exports = connectionFromMongoCursor;