+++ src/connectionFromMongooseAggregate.js
@@ -1,26 +1,20 @@
-'use strict';
+const { getOffsetsFromArgs, getConnectionFromSlice } = require('./utils');
 
-var utils = require('./utils');
-var getOffsetsFromArgs = utils.getOffsetsFromArgs;
-var getConnectionFromSlice = utils.getConnectionFromSlice;
-
-function cloneAggregation(aggr) {
-  /* eslint-disable no-underscore-dangle */
-  var model = aggr._model.model(aggr._model.modelName);
+const cloneAggregation = (aggr) => {
+  const model = aggr._model.model(aggr._model.modelName);
   return model.aggregate(aggr._pipeline);
-  /* eslint-enable no-underscore-dangle */
-}
+};
 
-function connectionFromMongooseAggregate(aggr, inArgs, mapper) {
-  var args = inArgs || {};
-  var countAggr = cloneAggregation(aggr);
+const connectionFromMongooseAggregate = (aggr, inArgs, mapper) => {
+  const args = inArgs || {};
+  const countAggr = cloneAggregation(aggr);
 
   return countAggr
     .group({ _id: null, count: { $sum: 1 } })
     .exec()
-    .then(function countPromise(countArr) {
-      var count = countArr.length > 0 && countArr[0].count ? countArr[0].count : 0;
-      var pagination = getOffsetsFromArgs(args, count);
+    .then((countArr) => {
+      const count = countArr.length > 0 && countArr[0].count ? countArr[0].count : 0;
+      const pagination = getOffsetsFromArgs(args, count);
 
       if (pagination.limit <= 0) {
         return getConnectionFromSlice([], mapper, args, count);
@@ -30,10 +24,10 @@ function connectionFromMongooseAggregate(aggr, inArgs, mapper) {
       aggr.limit(pagination.limit);
 
       return aggr.exec()
-        .then(function fromSlice(slice) {
+        .then((slice) => {
           return getConnectionFromSlice(slice, mapper, args, count);
         });
     });
-}
+};
 
 module.exports = connectionFromMongooseAggregate;